Restored naming and structure of repositories
