+++ parley/src/main/java/nu/parley/android/data/repository/MessageRepository.kt
@@ -1,23 +1,160 @@
-package nu.parley.android.data.repository;
+package nu.parley.android.data.repository
 
-import androidx.annotation.Nullable;
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+import nu.parley.android.Parley
+import nu.parley.android.data.model.Media
+import nu.parley.android.data.model.Message
+import nu.parley.android.data.model.MimeType.Companion.fromValue
+import nu.parley.android.data.net.ParleyHttpRequestMethod
+import nu.parley.android.data.net.RepositoryCallback
+import nu.parley.android.data.net.response.ParleyPaging
+import nu.parley.android.data.net.response.ParleyResponse
+import nu.parley.android.data.net.response.ParleyResponsePostMedia
+import nu.parley.android.data.net.response.ParleyResponsePostMessage
+import nu.parley.android.util.FileUtil
+import java.io.File
 
-import java.util.List;
+final class MessageRepository {
+    private class MessageResponseTypeToken : TypeToken<ParleyResponse<Message>>()
+    private class MessagesResponseTypeToken : TypeToken<ParleyResponse<List<Message>>>()
+    private class MessagesResponsePostMessageTypeToken : TypeToken<ParleyResponse<ParleyResponsePostMessage>>()
+    private class MessagesResponsePostMediaTypeToken : TypeToken<ParleyResponse<ParleyResponsePostMedia>>()
 
-import nu.parley.android.data.model.Message;
-import nu.parley.android.data.net.RepositoryCallback;
-import nu.parley.android.data.net.response.ParleyPaging;
-import nu.parley.android.data.net.response.ParleyResponse;
+    public fun findAll(callback: RepositoryCallback<ParleyResponse<List<Message>>>) {
+        val network = Parley.getInstance().network
+        network.networkSession.request(
+            network.url + network.path + "messages",
+            null,
+            ParleyHttpRequestMethod.Get,
+            emptyMap(),
+            onCompetion = {
+                val response = Gson()
+                    .getAdapter(MessagesResponseTypeToken())
+                    .fromJson(it)
+                callback.onSuccess(response)
+            },
+            onFailed = { statusCode, message ->
+                callback.onFailed(statusCode, message)
+            }
+        )
+    }
 
-public interface MessageRepository {
+    public fun getOlder(
+        previousPaging: ParleyPaging,
+        callback: RepositoryCallback<ParleyResponse<List<Message>>>
+    ) {
+        val network = Parley.getInstance().network
+        network.networkSession.request(
+            network.url + network.path + previousPaging.before,
+            null,
+            ParleyHttpRequestMethod.Get,
+            emptyMap(),
+            onCompetion = {
+                val response = Gson()
+                    .getAdapter(MessagesResponseTypeToken())
+                    .fromJson(it)
+                callback.onSuccess(response)
+            },
+            onFailed = { statusCode, message ->
+                callback.onFailed(statusCode, message)
+            }
+        )
+    }
 
-    public void findAll(final RepositoryCallback<ParleyResponse<List<Message>>> callback);
+    public fun send(
+        message: Message,
+        media: String?,
+        callback: RepositoryCallback<Message>
+    ) {
+        val network = Parley.getInstance().network
+        val onCompetion: (String) -> Unit = {
+            val response = Gson()
+                .getAdapter(MessagesResponsePostMessageTypeToken())
+                .fromJson(it)
 
-    public void getOlder(final ParleyPaging previousPaging, final RepositoryCallback<ParleyResponse<List<Message>>> callback);
+            var updatedMessage = Message.withIdAndStatus(
+                message,
+                response.data.messageId, Message.SEND_STATUS_SUCCESS
+            )
+            callback.onSuccess(updatedMessage)
+        }
+        val onFailed: (Int?, String?) -> Unit = { statusCode, errorMessage ->
+            callback.onFailed(statusCode, errorMessage)
+        }
+        val url = network.url + network.path + "messages"
+        if (media == null) {
+            network.networkSession.request(
+                url,
+                Gson().toJson(message),
+                ParleyHttpRequestMethod.Post,
+                emptyMap(),
+                onCompetion = onCompetion,
+                onFailed = onFailed
+            )
+        } else {
+            network.networkSession.upload(
+                url,
+                File(media),
+                FileUtil.getMimeType(File(media).absolutePath),
+                "image",
+                emptyMap(),
+                onCompetion = onCompetion,
+                onFailed = onFailed
+            )
+        }
+    }
 
-    public void send(final Message message, @Nullable final String media, final RepositoryCallback<Message> callback);
+    public fun sendMedia(
+        message: Message,
+        media: String,
+        callback: RepositoryCallback<Message>
+    ) {
+        val network = Parley.getInstance().network
+        val url = network.url + network.path + "media"
+        network.networkSession.upload(
+            url,
+            media,
+            FileUtil.getMimeType(media),
+            "media",
+            emptyMap(),
+            onCompetion = {
+                val response = Gson()
+                    .getAdapter(MessagesResponsePostMediaTypeToken())
+                    .fromJson(it)
 
-    public void sendMedia(final Message message, final String media, final RepositoryCallback<Message> callback);
+                val mediaType = FileUtil.getMimeType(media)
+                val mimeType = fromValue(mediaType!!)
+                val updatedMessage = Message.withMedia(
+                    message,
+                    Media(
+                        response.data.mediaId!!, File(media).name, mimeType.value
+                    )
+                )
+                callback.onSuccess(updatedMessage)
+            },
+            onFailed = { statusCode, errorMessage ->
+                callback.onFailed(statusCode, errorMessage)
+            }
+        )
+    }
 
-    public void get(final Integer messageId, final RepositoryCallback<Message> callback);
+    public fun get(messageId: Int, callback: RepositoryCallback<Message>) {
+        val network = Parley.getInstance().network
+        network.networkSession.request(
+            network.url + network.path + "messages/$messageId",
+            null,
+            ParleyHttpRequestMethod.Get,
+            emptyMap(),
+            onCompetion = {
+                val response = Gson()
+                    .getAdapter(MessageResponseTypeToken())
+                    .fromJson(it)
+                callback.onSuccess(response.data)
+            },
+            onFailed = { statusCode, message ->
+                callback.onFailed(statusCode, message)
+            }
+        )
+    }
 }
