+++ parley/src/main/java/nu/parley/android/view/compose/ParleyComposeListener.java
@@ -1,12 +1,12 @@
 package nu.parley.android.view.compose;
 
+import static nu.parley.android.notification.PushNotificationHandler.EVENT_START_TYPING;
+import static nu.parley.android.notification.PushNotificationHandler.EVENT_STOP_TYPING;
+
 import java.io.File;
 
 import nu.parley.android.Parley;
-import nu.parley.android.data.repository.DefaultEventRepository;
-
-import static nu.parley.android.notification.PushNotificationHandler.EVENT_START_TYPING;
-import static nu.parley.android.notification.PushNotificationHandler.EVENT_STOP_TYPING;
+import nu.parley.android.data.repository.EventRepository;
 
 public final class ParleyComposeListener implements ComposeListener {
 
@@ -31,6 +31,6 @@ public final class ParleyComposeListener implements ComposeListener {
     }
 
     private void fireEvent(String event) {
-        new DefaultEventRepository().fire(event);
+        new EventRepository().fire(event);
     }
 }
